+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.8.0
+
+- When twisted is installed, emit() returns a deferred which gathers the
+  results of all callbacks.
+
 ## 0.7.0
 
 - Fixed #3: Added disconnect/disconnect_from partials to callbacks